+++ src/apis/types/feed.ts
@@ -91,56 +91,3 @@ export interface FeedDetail {
   interests: string[];
   series: Record<string, string>;
 }
-
-export interface GetFeedCommentsResponse extends CommentResponse {
-  isFeedWriter: boolean;
-}
-
-export interface PostFeedCommentParams {
-  feedId: number;
-  comment: string;
-}
-
-export interface PostFeedCommentResponse {
-  userId: number;
-  userName: string;
-  role: Role;
-  isWriter: boolean;
-  commentId: number;
-  comment: string;
-  createdAt: string;
-}
-
-export interface PostFeedCommentReplyParams {
-  feedId: number;
-  comment: string;
-  parentCommentId: number;
-}
-
-export interface PostFeedCommentReplyResponse {
-  userId: number;
-  userName: string;
-  role: Role;
-  isWriter: boolean;
-  commentId: number;
-  comment: string;
-  createdAt: string;
-  parentCommentId: number;
-}
-
-export interface PatchFeedCommentParams {
-  commentId: number;
-  comment: string;
-}
-
-export interface CommentResponse {
-  userId: number;
-  userName: string;
-  role: Role;
-  isCommentWriter: boolean;
-  commentId: number;
-  comment: string;
-  parentCommentId: number;
-  createdAt: string;
-  children: CommentResponse[];
-}
